Refactor PNG processing by extracting utility functions for image processing and indexed data extraction
